Update to MonoGame 3.6 Addresses prime31/Nez#220

+++ Nez.Portable/Core.cs
@@ -313,7 +313,7 @@ namespace Nez
 				TimeRuler.instance.render();
 
 			#if !FNA
-			drawCalls = graphicsDevice.Metrics.DrawCount;
+			drawCalls = (ulong)graphicsDevice.Metrics.DrawCount;
 			#endif
 			#endif
 		}
